requirements.txt
networkx
numpy
csv

pandas
matplotlib

pytorch
tensorboard

gym
stable_baselines3

pytorch_geometric